Middlewares, new file, and examples
